refactor(tuner,ui): shared early-stop helper; drop unreachable FINISHED view
- _should_early_stop was duplicated 1:1 between the gmx and amber
  runners and already drifting; extract it to engines/early_stop.py.
- The dashboard auto-stops tuner jobs on FINISHED, so tuner_status
  never reaches the UI as FINISHED — the displayFinished/jobFinished
  branches were unreachable. The stopped view is the same persistence
  surface (preserved measured trials + Delete); remove the dead code.
