+++ dashboard/ui/src/components/Wizard/TuneStep/TunerView.tsx
@@ -59,8 +59,6 @@ const TunerView = (props: TunerViewProps) => {
 
   const displayStarted = tunerStarted_condition(tuner)
   const displayStopped = tuner?.is_stopped || false
-  // FINISHED is a terminal tuner-side status: nothing left to stop, only delete.
-  const displayFinished = !displayStopped && tuner?.tuner_status === "FINISHED"
   const trials = (tuner?.trials || []) as GmxTunerTrial[]
   const unavailableReason = simulationLaunchUnavailableReason(simulation ?? null, experiment.engine)
 
@@ -73,12 +71,11 @@ const TunerView = (props: TunerViewProps) => {
             selectedTrial={selectedTrial}
             setSelectedTrial={setSelectedTrial}
             tunerStopped={displayStopped}
-            jobFinished={displayFinished}
             experimentId={experiment.id}
             simulationPath={simulationPath}
           />
 
-          {!displayStopped && !displayFinished && (
+          {!displayStopped && (
             <div className="flex justify-end gap-2">
               <Button
                 variant="default"
@@ -91,7 +88,7 @@ const TunerView = (props: TunerViewProps) => {
             </div>
           )}
 
-          {(displayStopped || displayFinished) && (
+          {displayStopped && (
             <div className="flex justify-end gap-2">
               <Button
                 variant="outline"
